requirements.txt
streamlit==1.28.0
streamlit-chat==0.1.1
python-dotenv==1.0.0
google-generativeai==0.3.0
